perf(attributes): add a new system for tracking attributes in blender or typing them manually
Change the current attribute handling system to a more advanced yaml format, which gives the
possibility of showing all GE attributes in blender, for a nice overview. But also locking some of
them to follow their blender counterpart. For the ultimate freedom of how you want to handle the
attributes