feat(export): Add Option to Binarize i3d file for export
This adds an option to specify a path to the `i3dConverter.exe`, which can then be used to create a binary i3d file. The converter exe is NOT bundled with the release and has to be supplied by the user. It can be found as part of the official Giants I3D exporter, which is downloadable from https://gdn.giants-software.com/downloads.php.