feat(skinned-mesh): add support for 'Child Of' constraint on bones
Make it possible for bones in armatures to set a custom parent
based on the 'Child Of' constraint target (PoseBone). Simplifies
the workflow by allowing a single armature to manage
all bones, instead of requiring one armature per bone.

Note: With current implementation the armature need to be added outside the hierarchy at the bottom of everything you want to export