fix(attributes): update default values for Cast and Receive Shadows (#243)
* fix(attributes): update default values for Cast and Receive Shadows

Add a new `blender_default` value to the `i3d_map` to allow for a different default in Blender compared with the normal i3d default value. Change `casts_shadows` and `receive_shadows` to `True` in Blender, aligning better with typical use cases where shadows are expected. It's generally easier to disable shadows for a few collisions than to detect missing shadows on meshes