fix(exporter): skip invalid export settings
If a user opens a .blend file that was saved with a newer version of the exporter (with additional export settings), some of those settings might not exist in the older version they have installed.
This situation is rare, but it can happen when sharing files or switching back to a previous release.