fix(mesh): remove "use vertex colors" property (#295)
Removed the "use vertex colors" property after discussions across multiple channels. We concluded that the property is likely to cause more confusion than benefit. It's difficult to expose this setting in a clearly visible way, which could lead to misunderstandings—especially for new users who expect vertex colors to be exported by default.

We may revisit this in the future with a better approach.